Documents/requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
pillow
opencv-python
pydantic
ultralytics
python-multipart